refactor: Merge abstract widget into metadata widget
