+++ projects/hslayers/src/components/layermanager/editor/layer-editor.component.ts
@@ -6,7 +6,6 @@ import {Geometry} from 'ol/geom';
 import {Layer} from 'ol/layer';
 import {Source} from 'ol/source';
 
-import {HsAbstractWidgetComponent} from '../widgets/abstract-widget.component';
 import {HsConfirmDialogComponent} from './../../../common/confirm/confirm-dialog.component';
 import {HsDialogContainerService} from '../../layout/dialogs/dialog-container.service';
 import {HsDimensionTimeService} from '../../../common/get-capabilities/dimension-time.service';
@@ -74,10 +73,6 @@ export class HsLayerEditorComponent {
     this.hsLayerEditorWidgetContainerService.create(HsTypeWidgetComponent, {});
     this.hsLayerEditorWidgetContainerService.create(
       HsMetadataWidgetComponent,
-      {}
-    );
-    this.hsLayerEditorWidgetContainerService.create(
-      HsAbstractWidgetComponent,
       {}
     );
     this.hsLayerEditorWidgetContainerService.create(HsScaleWidgetComponent, {});
