--- projects/hslayers/src/components/layermanager/widgets/abstract-widget.component.html
@@ -1,6 +1,0 @@
-<div class="form-group" [hidden]="!abstract">
-  <label>{{ "COMMON.abstract" | translate }}</label>
-  <p
-    [innerHtml]="HsLayerManagerService.makeSafeAndTranslate('LAYERS', abstract)"
-  ></p>
-</div>
